+++ backend/fittoring/src/main/java/fittoring/mentoring/business/service/MentoringReservationService.java
@@ -2,6 +2,8 @@ package fittoring.mentoring.business.service;
 
 import fittoring.mentoring.business.service.dto.ReservationCreateDto;
 import fittoring.mentoring.business.service.dto.SmsReservationMessageDto;
+import fittoring.mentoring.infra.SmsMessageFormatter;
+import fittoring.mentoring.infra.SmsRestClientService;
 import fittoring.mentoring.presentation.dto.ReservationCreateResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -10,6 +12,8 @@ import org.springframework.stereotype.Service;
 @Service
 public class MentoringReservationService {
 
+    private static final String RESERVATION_SUBJECT = "핏토링 예약 알림";
+
     private final ReservationService reservationService;
     private final SmsRestClientService smsRestClientService;
     private final SmsMessageFormatter smsMessageFormatter;
@@ -17,7 +21,7 @@ public class MentoringReservationService {
     public ReservationCreateResponse reserveMentoring(ReservationCreateDto dto) {
         ReservationCreateResponse response = reservationService.createReservation(dto);
         String smsMessage = smsMessageFormatter.createSmsReservationMessage(SmsReservationMessageDto.of(dto));
-        smsRestClientService.sendSms(response.mentorPhone(), smsMessage);
+        smsRestClientService.sendSms(response.mentorPhone(), RESERVATION_SUBJECT, smsMessage);
         return response;
     }
 }
