+++ backend/fittoring/src/main/java/fittoring/mentoring/infra/SmsMessageFormatter.java
@@ -1,4 +1,4 @@
-package fittoring.mentoring.business.service;
+package fittoring.mentoring.infra;
 
 import fittoring.mentoring.business.service.dto.SmsReservationMessageDto;
 import org.springframework.stereotype.Component;
